+++ test/regression/inrprod1.cpp
@@ -19,7 +19,7 @@ int inrprod1_test(int, char**)
   cout<<"Results of inrprod1_test:"<<endl;
   vector <int> v1(3);
   vector <int> v2(v1.size());
-  for(int i = 0; i < v1.size(); i++)
+  for(size_t i = 0; i < v1.size(); i++)
   {
     v1[i] = i + 1;
     v2[i] = v1.size() - i;
